proposed wording change in line 59
just a small update to the wording. I find that saying content that displays on the webpage "should not" be not be in the header still sounds like it is more a bad practice rather than that information in the head is not directly displays. A small thing but I was thinking it might be clearer for someone who is just starting.